Добавлен хук useSearch.js Исправлен баг: клик по ссылке поиска
