+++ src/components/Search/Search.jsx
@@ -1,10 +1,11 @@
-import { useState, useContext } from "react";
+import { useContext } from "react";
 import { useDispatch, useSelector } from "react-redux";
 import { useNavigate } from "react-router";
 import { DebounceInput } from "react-debounce-input";
 
 import { useGetCharacterByNameQuery } from "../../features/api/apiSlice";
 import { addToHistory, selectLoggedIn } from "../../features";
+import { useSearch } from "../../hooks/useSearch";
 import { SearchItem } from "./SearchItem";
 
 import { ThemeContext } from "../../app/themeContext";
@@ -15,24 +16,13 @@ import classes from "./styles/searchStyle.module.css";
 export function Search() {
   const { theme } = useContext(ThemeContext);
 
-  const [searchValue, setSearchValue] = useState("");
-
-  const [isVisibleSearchField, setSearchFieldVisible] = useState(false);
+  const { hideSearchField, search, isVisibleSearchField, searchValue } = useSearch();
 
   const isLoggedIn = useSelector(selectLoggedIn);
 
   const navigate = useNavigate();
 
   const dispatch = useDispatch();
-
-  const hideSearchField = () => {
-    setSearchFieldVisible(false);
-  };
-
-  const search = (newValue) => {
-    setSearchValue(newValue);
-    setSearchFieldVisible(newValue !== "");
-  };
 
   const {
     data = [],
@@ -41,12 +31,16 @@ export function Search() {
     isSuccess,
   } = useGetCharacterByNameQuery(searchValue);
 
+  const hideAndPushToHistory = () => {
+    hideSearchField();
+    if (isLoggedIn) {
+      dispatch(addToHistory(searchValue));
+    }
+  }
+
   const onKeyPress = (event) => {
     if (event.key === "Enter" && searchValue !== "") {
-      hideSearchField();
-      if (isLoggedIn) {
-        dispatch(addToHistory(searchValue));
-      }
+      hideAndPushToHistory();
       navigate(`/search/${searchValue}`);
     }
   };
@@ -62,7 +56,7 @@ export function Search() {
       <DebounceInput
         onKeyDown={onKeyPress}
         minLength={2}
-        debounceTimeout={500}
+        debounceTimeout={200}
         placeholder="Панель поиска"
         className={classes.search}
         onChange={(event) => search(event.target.value)}
@@ -84,7 +78,7 @@ export function Search() {
                 key={card.id}
                 id={card.id}
                 name={card.name}
-                hideleSearchFieldVisible={hideSearchField}
+                linkClickHandler={hideAndPushToHistory}
               />
             ))}
         </nav>
