+++ src/components/Search/SearchItem/SearchItem.jsx
@@ -1,26 +1,17 @@
 import PropTypes from "prop-types";
 import { Link } from "react-router-dom";
-import { useDispatch } from "react-redux";
-
-import { addToHistory } from "../../../features";
 import { chooseThemeClass } from "../../../utils";
-import { CardButton } from "../../Card/cardButton";
+
+import CardButton from "../../Card/cardButton/CardButton";
 
 import classes from "./styles/searchItemStyles.module.css";
 
-export function SearchItem({ theme, id, name, hideSearchField }) {
-  const dispatch = useDispatch();
-
-  const onLinkClick = () => {
-    dispatch(addToHistory(name));
-    console.log(name);
-    hideSearchField();
-  };
+export function SearchItem({ theme, id, name, linkClickHandler }) {
 
   return (
     <div className={`${chooseThemeClass(theme, classes.searchItem_dark)} ${classes.searchItem}`}>
       <Link
-        onClick={onLinkClick}
+        onClick={linkClickHandler}
         to={`/search/${name}`}
         className={classes.searchItem__name}
       >
@@ -36,5 +27,5 @@ SearchItem.propTypes = {
   theme: PropTypes.string,
   id: PropTypes.number,
   name: PropTypes.string,
-  hideSearchField: PropTypes.bool
+  hideSearchField: PropTypes.func
 };
